requirements.txt
flask==3.0.0
redis==5.0.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
requests==2.31.0
flask-cors==4.0.0
fakeredis==2.20.1

